# api/openapi/swagger.yaml
basePath: /api
definitions:
...
  handler.TenantListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/handler.TenantResponse'
        type: array
      total:
        type: integer
    type: object
  handler.TenantResponse:
    properties:
      created_at:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: integer
    type: object
...
  response.errorResponse:
    properties:
      code:
        type: integer
      details:
        additionalProperties: true
        type: object
      message:
        example: Internal server error
        type: string
    type: object
  response.invalidParamsResponse:
    properties:
      code:
        type: integer
      details:
        additionalProperties: true
        type: object
      message:
        example: invalid params
        type: string
    type: object
  response.successResponse:
    properties:
      code:
        example: 2000
        type: integer
      data: {}
      message:
        example: 请求成功
        type: string
    type: object
...
  saas_internal_tenant_handler.CreateRequest:
    properties:
      description:
        maxLength: 120
        type: string
      name:
        type: string
      plan_id:
        type: integer
    required:
    - name
    - plan_id
    type: object
  saas_internal_tenant_handler.UpdateRequest:
    properties:
      description:
        maxLength: 120
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: lirous@lirous.com
    name: Lirous
    url: https://Lirous.com
  description: 自定义描述
  license:
    name: MIT
    url: https://github.com/Lirous587/go-scaffold/main/LICENSE
  termsOfService: https://Lirous.com
  title: 自定义title
  version: "1.0"
paths:
...
  /v1/tenant:
    get:
      consumes:
      - application/json
      description: 根据查询参数获取Tenant列表，返回当前页数据和total数量
      parameters:
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Tenant列表
          schema:
            allOf:
            - $ref: '#/definitions/response.successResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.TenantListResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.invalidParamsResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.errorResponse'
      summary: 获取 Tenant 列表
      tags:
      - tenant
    post:
      consumes:
      - application/json
      description: 创建新的 Tenant
      parameters:
      - description: 创建 Tenant 请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/saas_internal_tenant_handler.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功创建 Tenant
          schema:
            allOf:
            - $ref: '#/definitions/response.successResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.TenantResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.invalidParamsResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.errorResponse'
      security:
      - BearerAuth: []
      summary: 用户新建租户
      tags:
      - tenant
  /v1/tenant/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除 Tenant
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除 Tenant
          schema:
            $ref: '#/definitions/response.successResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.invalidParamsResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.errorResponse'
      security:
      - BearerAuth: []
      summary: 删除 Tenant
      tags:
      - tenant
    get:
      consumes:
      - application/json
      description: 读取单条 Tenant
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功创建 Tenant
          schema:
            allOf:
            - $ref: '#/definitions/response.successResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.TenantResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.invalidParamsResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.errorResponse'
      security:
      - BearerAuth: []
      summary: 读取单条 Tenant
      tags:
      - tenant
    put:
      consumes:
      - application/json
      description: 根据ID更新 Tenant 信息
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新 Tenant 请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/saas_internal_tenant_handler.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功更新 Tenant
          schema:
            allOf:
            - $ref: '#/definitions/response.successResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.TenantResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.invalidParamsResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.errorResponse'
      security:
      - BearerAuth: []
      summary: 更新 Tenant
      tags:
      - tenant
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"