# internal/tenant/handler/handler.go
package handler

import (
	"github.com/gin-gonic/gin"
	"github.com/pkg/errors"
	"saas/internal/common/reqkit/bind"
	"saas/internal/common/reskit/response"
	"saas/internal/common/server"
	"saas/internal/tenant/domain"
	"strconv"
)

type HttpHandler struct {
	service domain.TenantService
}

func NewHttpHandler(service domain.TenantService) *HttpHandler {
	return &HttpHandler{
		service: service,
	}
}

func (h *HttpHandler) getID(ctx *gin.Context) (int64, error) {
	idStr := ctx.Param("id")
	idInt, err := strconv.Atoi(idStr)
	if err != nil {
		response.InvalidParams(ctx, err)
		return 0, err
	}
	if idInt == 0 {
		err := errors.New("无效的id")
		response.InvalidParams(ctx, err)
		return 0, err
	}
	return int64(idInt), err
}

// Create godoc
// @Summary      用户新建租户
// @Description  创建新的 Tenant
// @Tags         tenant
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        request body handler.CreateRequest true "创建 Tenant 请求"
// @Success      200  {object}  response.successResponse{data=handler.TenantResponse} "成功创建 Tenant"
// @Failure      400  {object}  response.invalidParamsResponse "参数错误"
// @Failure      500  {object}  response.errorResponse "服务器错误"
// @Router       /v1/tenant [post]
func (h *HttpHandler) Create(ctx *gin.Context) {
	req := new(CreateRequest)

	userID, err := server.GetUserID(ctx)
	if err != nil {
		response.InvalidParams(ctx, err)
		return
	}

	if err := bind.BindingRegularAndResponse(ctx, req); err != nil {
		return
	}

	data, err := h.service.Create(&domain.Tenant{
		Name:        req.Name,
		Description: req.Description,
	},
		req.PlanID,
		userID,
	)

	if err != nil {
		response.Error(ctx, err)
		return
	}

	response.Success(ctx, domainTenantToResponse(data))
}

// Read godoc
// @Summary      读取单条 Tenant
// @Description  读取单条 Tenant
// @Tags         tenant
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        id   path int true "Tenant ID"
// @Success      200  {object}  response.successResponse{data=handler.TenantResponse} "成功创建 Tenant"
// @Failure      400  {object}  response.invalidParamsResponse "参数错误"
// @Failure      500  {object}  response.errorResponse "服务器错误"
// @Router       /v1/tenant/{id} [get]
func (h *HttpHandler) Read(ctx *gin.Context) {
	id, err := h.getID(ctx)
	if err != nil {
		return
	}

	data, err := h.service.Read(id)

	if err != nil {
		response.Error(ctx, err)
		return
	}

	response.Success(ctx, domainTenantToResponse(data))
}

// Update godoc
// @Summary      更新 Tenant
// @Description  根据ID更新 Tenant 信息
// @Tags         tenant
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        id   path int true "Tenant ID"
// @Param        request body handler.UpdateRequest true "更新 Tenant 请求"
// @Success      200  {object}  response.successResponse{data=handler.TenantResponse} "成功更新 Tenant"
// @Failure      400  {object}  response.invalidParamsResponse "参数错误"
// @Failure      500  {object}  response.errorResponse "服务器错误"
// @Router       /v1/tenant/{id} [put]
func (h *HttpHandler) Update(ctx *gin.Context) {
	id, err := h.getID(ctx)
	if err != nil {
		return
	}

	req := new(UpdateRequest)

	if err := bind.BindingRegularAndResponse(ctx, req); err != nil {
		return
	}

	data, err := h.service.Update(&domain.Tenant{
		ID:          id,
		Name:        req.Name,
		Description: req.Description,
	})

	if err != nil {
		response.Error(ctx, err)
		return
	}

	response.Success(ctx, domainTenantToResponse(data))
}

// Delete godoc
// @Summary      删除 Tenant
// @Description  根据ID删除 Tenant
// @Tags         tenant
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        id   path int true "Tenant ID"
// @Success      200  {object}  response.successResponse "成功删除 Tenant"
// @Failure      400  {object}  response.invalidParamsResponse "参数错误"
// @Failure      500  {object}  response.errorResponse "服务器错误"
// @Router       /v1/tenant/{id} [delete]
func (h *HttpHandler) Delete(ctx *gin.Context) {
	id, err := h.getID(ctx)
	if err != nil {
		return
	}

	if err := h.service.Delete(id); err != nil {
		response.Error(ctx, err)
		return
	}

	response.Success(ctx)
}

// List godoc
// @Summary      获取 Tenant 列表
// @Description  根据查询参数获取Tenant列表，返回当前页数据和total数量
// @Tags         tenant
// @Accept       json
// @Produce      json
// @Param        keyword    query     string  false  "关键词搜索"
// @Param        page       query     int     false  "页码" default(1)
// @Param        page_size  query     int     false  "每页数量" default(10)
// @Success      200  {object}  response.successResponse{data=handler.TenantListResponse} "Tenant列表"
// @Failure      400  {object}  response.invalidParamsResponse "参数错误"
// @Failure      500  {object}  response.errorResponse "服务器错误"
// @Router       /v1/tenant [get]
func (h *HttpHandler) List(ctx *gin.Context) {
	req := new(ListRequest)

	if err := bind.BindingRegularAndResponse(ctx, req); err != nil {
		return
	}

	data, err := h.service.List(&domain.TenantQuery{
		Keyword:  req.KeyWord,
		Page:     req.Page,
		PageSize: req.PageSize,
	})

	if err != nil {
		response.Error(ctx, err)
		return
	}

	response.Success(ctx, domainTenantListToResponse(data))
}